removed unnecessary api request
