--- .storybook/preview.js
@@ -1,16 +1,0 @@
-import '../src/index.scss';
-
-/** @type { import('@storybook/react').Preview } */
-const preview = {
-	parameters: {
-		actions: { argTypesRegex: '^on[A-Z].*' },
-		controls: {
-			matchers: {
-				color: /(background|color)$/i,
-				date: /Date$/,
-			},
-		},
-	},
-};
-
-export default preview;
