+++ package.json
@@ -1,110 +1,52 @@
 {
-	"name": "storybook-example",
-	"version": "0.1.0",
-	"private": true,
-	"dependencies": {
-		"@testing-library/jest-dom": "^5.14.1",
-		"@testing-library/react": "^13.0.0",
-		"@testing-library/user-event": "^13.2.1",
-		"bulma": "^0.9.4",
-		"clsx": "^1.2.1",
-		"prop-types": "15.8.1",
-		"react": "^18.2.0",
-		"react-dom": "^18.2.0",
-		"react-scripts": "5.0.1",
-		"sass": "^1.62.1",
-		"web-vitals": "^2.1.0"
-	},
-	"scripts": {
-		"start": "react-scripts start",
-		"build": "react-scripts build",
-		"test": "react-scripts test",
-		"eject": "react-scripts eject",
-		"storybook": "storybook dev -p 6006",
-		"build-storybook": "storybook build",
-		"lint": "eslint . --ext .js,.jsx .",
-		"lint:fix": "npm run lint -- --fix",
-		"format": "prettier . --write",
-		"prepare": "husky install",
-		"commit": "cz"
-	},
-	"eslintConfig": {
-		"extends": [
-			"react-app",
-			"react-app/jest",
-			"airbnb",
-			"airbnb/hooks",
-			"prettier"
-		],
-		"rules": {
-			"react/jsx-filename-extension": [
-				1,
-				{
-					"extensions": [
-						".js",
-						".jsx"
-					]
-				}
-			],
-			"react/react-in-jsx-scope": "off",
-			"import/prefer-default-export": "off",
-			"react/function-component-definition": "off",
-			"react/jsx-props-no-spreading": "off",
-			"react/no-unescaped-entities": "off",
-			"react/button-has-type": "off"
-		},
-		"overrides": [
-			{
-				"files": [
-					"**/*.stories.*",
-					"src/stories/**/*"
-				],
-				"rules": {
-					"import/no-anonymous-default-export": "off",
-					"import/no-extraneous-dependencies": "off"
-				}
-			}
-		]
-	},
-	"browserslist": {
-		"production": [
-			">0.2%",
-			"not dead",
-			"not op_mini all"
-		],
-		"development": [
-			"last 1 chrome version",
-			"last 1 firefox version",
-			"last 1 safari version"
-		]
-	},
-	"devDependencies": {
-		"@storybook/addon-essentials": "7.0.7",
-		"@storybook/addon-interactions": "7.0.7",
-		"@storybook/addon-links": "7.0.7",
-		"@storybook/blocks": "7.0.7",
-		"@storybook/preset-create-react-app": "7.0.7",
-		"@storybook/react": "7.0.7",
-		"@storybook/react-webpack5": "7.0.7",
-		"@storybook/testing-library": "0.0.14-next.2",
-		"babel-plugin-named-exports-order": "0.0.2",
-		"commitizen": "^4.3.0",
-		"cz-conventional-changelog": "^3.3.0",
-		"eslint-config-airbnb": "^19.0.4",
-		"eslint-config-prettier": "^8.8.0",
-		"husky": "^8.0.3",
-		"lint-staged": "^13.2.1",
-		"prettier": "^2.8.8",
-		"storybook": "7.0.7",
-		"webpack": "5.81.0"
-	},
-	"lint-staged": {
-		"*.{js,jsx}": "npm run lint",
-		"*.{js,jsx,css,md,html,json}": "npm run format"
-	},
-	"config": {
-		"commitizen": {
-			"path": "./node_modules/cz-conventional-changelog"
-		}
-	}
+  "name": "linguista_web",
+  "version": "0.1.0",
+  "private": true,
+  "dependencies": {
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.15.0",
+    "react-scripts": "5.0.1",
+    "web-vitals": "^2.1.4"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject",
+    "lint": "eslint . --ext .js,.jsx .",
+    "lint:fix": "npm run lint -- --fix",
+    "format": "prettier . --write"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest",
+      "prettier"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  },
+  "devDependencies": {
+    "babel-eslint": "^10.1.0",
+    "eslint": "^8.48.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-prettier": "^5.0.0",
+    "eslint-plugin-react": "^7.33.2",
+    "node-sass": "^7.0.3",
+    "prettier": "3.0.0"
+  }
 }
