+++ src/components/App/App.module.scss
@@ -2,18 +2,18 @@
 	text-align: center;
 }
 
-.App-logo {
+.App_logo {
 	height: 40vmin;
 	pointer-events: none;
 }
 
 @media (prefers-reduced-motion: no-preference) {
-	.App-logo {
+	.App_logo {
 		animation: App-logo-spin infinite 20s linear;
 	}
 }
 
-.App-header {
+.App_header {
 	background-color: #282c34;
 	min-height: 100vh;
 	display: flex;
@@ -24,15 +24,7 @@
 	color: white;
 }
 
-.App-link {
+.App_link {
 	color: #61dafb;
 }
 
-@keyframes App-logo-spin {
-	from {
-		transform: rotate(0deg);
-	}
-	to {
-		transform: rotate(360deg);
-	}
-}
