+++ src/styles/index.scss
@@ -1,4 +1,3 @@
-@import '~bulma';
 
 body {
 	margin: 0;
